+++ CHANGELOG.md
@@ -79,6 +79,7 @@ Developers may need to change their code
 
 ### Corrected Bugs
 
+* Fixed error memory in RESTCONF PATCH/PUT when accessing top-level data node.
 * Fixed: [ Calling copy-config RPC from restconf #158](https://github.com/clicon/clixon/issues/158)
 * Fixed: [namespace prefix nc is not supported in full #154](https://github.com/clicon/clixon/issues/154)
   * edit-config "config" parameter did not work with prefix other than null